refactor activity tests into /activities
