+++ app/src/androidTest/java/com/github/fribourgsdp/radio/activities/EndGameActivityTest.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.activities
 
 import android.content.Context
 import android.content.Intent
@@ -10,6 +10,8 @@ import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.fribourgsdp.radio.MainActivity
+import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.game.EndGameActivity
 import com.github.fribourgsdp.radio.game.SCORES_KEY
 import com.github.fribourgsdp.radio.utils.CustomMatchers.Companion.atPosition
