+++ app/src/androidTest/java/com/github/fribourgsdp/radio/activities/WorkingJoinGameActivityTest.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.activities
 
 import android.Manifest
 import android.content.Context
@@ -14,6 +14,7 @@ import androidx.test.espresso.matcher.RootMatchers
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.GrantPermissionRule
+import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.game.prep.*
 import com.github.fribourgsdp.radio.mockimplementations.*
 import org.hamcrest.Matchers
